+++ include/cRIO/ElectromechanicalPneumaticILC.h
@@ -61,6 +61,7 @@ public:
     enum ILC_EM_CMD {
         SET_STEPPER_STEPS = 66,
         STEPPER_FORCE_STATUS = 67,
+        FREEZE_SENSOR = 68,
         SET_DCA_GAIN = 73,
         REPORT_DCA_GAIN = 74,
         SET_FORCE_OFFSET = 75,
@@ -70,6 +71,8 @@ public:
         REPORT_MEZZANINE_PRESSURE = 119,
         REPORT_HARDPOINT_LVDT = 122
     };
+
+    static constexpr uint8_t EA_BROADCAST = 248;
 
     /**
      * Unicast command to command stepper motor moves.
@@ -84,6 +87,19 @@ public:
      */
     void setStepperSteps(uint8_t address, int8_t steps) {
         callFunction(address, ILC_EM_CMD::SET_STEPPER_STEPS, 1800, steps);
+    }
+
+    /**
+     * Broadcast steps to all force actuators.
+     *
+     * @param counter broadcast counter (0-15)
+     * @param steps commanded steps
+     *
+     * @ingroup M1M3_hp
+     * @ingroup M2
+     */
+    void broadcastStepperSteps(uint8_t counter, std::vector<int8_t> steps) {
+        broadcastFunction(EA_BROADCAST, ILC_EM_CMD::SET_STEPPER_STEPS, 1800, counter, steps);
     }
 
     /**
@@ -167,6 +183,21 @@ public:
      */
     void reportForceActuatorForceStatus(uint8_t address) {
         callFunction(address, ILC_EM_CMD::REPORT_FA_FORCE_STATUS, 1800);
+    }
+
+    /**
+     * Freeze sensor values. After issuing freeze, sensor values can be read
+     * out with reportForceActuatorForceStatus (function
+     * REPORT_FA_FORCE_STATUS).
+     *
+     * @param counter broadcast counter (0-15)
+     *
+     * @ingroup M1M3_fa
+     * @ingroup M1M3_hp
+     * @ingroup M2
+     */
+    void freezeSensor(uint8_t counter) {
+        broadcastFunction(EA_BROADCAST, ILC_EM_CMD::FREEZE_SENSOR, 180, counter);
     }
 
     /**