+++ doc/version-history.rst
@@ -2,6 +2,16 @@
 Version History
 ###############
 
+
+v1.11.2
+-------
+* SensorMonitor ILC
+
+v1.11.1
+-------
+* broadcast commands to step and freeze
+* SimpleFPGACliApp class
+
 v1.11.0
 -------
 * Modbus::, ILC::ILCBusList classes. Communication code moved from ModbusBuffer